Fixed the problem of implementation of a templated class.
The implementation of a templated class must be compiled before the .cpp
file which includes it, or the compiler will complain that theres's an
undefined reference to that templated class.
Thus, I include the content of the source code to the end of the header
file. And renamed its extension to .tpp for better recognition.